add history deletion on history page

+++ classes/Device.php
@@ -48,7 +48,7 @@ class Device
         }
     }
 
-    
+
     public function loadDevice($uuid)
     {
         $device=null;
@@ -63,9 +63,9 @@ class Device
         }
         return $device;
     }
-    
 
-    public function deleteDevice($uuid)
+
+    public function deleteHistory($uuid)
     {
         //error_log("Deleteing device =".$uuid);
         if ($this->databaseConnection()) {
@@ -82,18 +82,27 @@ class Device
             $query_device = $this->db_connection->prepare ( 'delete FROM motion_last WHERE uuid = :uuid' );
             $query_device->bindValue ( ':uuid', $uuid, PDO::PARAM_STR );
             $query_device->execute ();
-            $query_device = $this->db_connection->prepare('delete FROM device WHERE uuid = :uuid');
-            $query_device->bindValue(':uuid', $uuid, PDO::PARAM_STR);
+            //$query_device = $this->db_connection->prepare('delete FROM device WHERE uuid = :uuid');
+            //$query_device->bindValue(':uuid', $uuid, PDO::PARAM_STR);
             $query_device->execute();
             $query_device = $this->db_connection->prepare('delete FROM alert_config WHERE uuid = :uuid');
             $query_device->bindValue(':uuid', $uuid, PDO::PARAM_STR);
             $query_device->execute();
-            exec ('curl --silent "http://'.SIP_USER.":".SIP_PASS."@".SIP_HOST.":".SIP_PORT."/showUsers.html?remove.".$uuid.'@sip.ibeyonde.com=on"');
-            
             //error_log("Error=".implode(",", $query_device->errorInfo()));
         }
     }
-    
+
+
+    public function deleteDevice($uuid)
+    {
+        //error_log("Deleteing device =".$uuid);
+        if ($this->databaseConnection()) {
+            $query_device = $this->db_connection->prepare('delete FROM device WHERE uuid = :uuid');
+            $query_device->bindValue(':uuid', $uuid, PDO::PARAM_STR);
+            //error_log("Error=".implode(",", $query_device->errorInfo()));
+        }
+        $this->deleteDevice($uuid);
+    }
 
     public function alert($state)
     {
@@ -105,7 +114,7 @@ class Device
             $query_device->bindValue(':state', $state, PDO::PARAM_INT);
             $query_device->bindValue(':uuid', $this->uuid, PDO::PARAM_STR);
             $query_device->execute();
-        } 
+        }
         return 1;
     }
 
@@ -125,7 +134,7 @@ class Device
         }
         return 1;
     }
-    
+
     public function isCapable($capability){
         if (strpos($this->capabilities, $capability) !== false){
             return True;
@@ -134,7 +143,7 @@ class Device
             return False;
         }
     }
-    
+
 
     public static function loadUserDevices($username)
     {
@@ -156,7 +165,7 @@ class Device
         }
         return $devices;
     }
-    
+
 
     public static function loadAllDevices()
     {
@@ -177,7 +186,7 @@ class Device
         }
         return $devices;
     }
-    
+
     public static function getDeviceOwner($uuid){
         $db_connection = new PDO ( 'mysql:host=' . DB_HOST . ';dbname=' . DB_NAME . ';charset=utf8', DB_USER, DB_PASS );
         // database query, getting all the info of the selected user
@@ -186,7 +195,7 @@ class Device
         $query_device->execute();
         return $query_device->fetch()[0];
     }
-    
+
     public static function getDeviceName($uuid){
         $db_connection = new PDO ( 'mysql:host=' . DB_HOST . ';dbname=' . DB_NAME . ';charset=utf8', DB_USER, DB_PASS );
         // database query, getting all the info of the selected user
@@ -203,6 +212,6 @@ class Device
         $query_device->execute();
         return $query_device->fetch()[0];
     }
-    
-    
+
+
 }
